Extract shared IcebergSnapshotHelper trait from 6 duplicated test methods
Six Iceberg test files each had their own copy of appendSnapshot with
slight variations. Four still used the old direct-write pattern that
breaks on multi-batch tests. Consolidated into IcebergSnapshotHelper
trait in IcebergTestSupport.scala with staging+move to a common data/
directory, optional partitionPath, and explicit schema parameter.
